Core/Spawn: Limit saved/stored respawn time.  - Limit to long maximum, for the compiled-for architecture.
Port From (https://github.com/TrinityCore/TrinityCore/commit/f3ac10f36fcc46c5e1537d69544a9294fc719ad2)

+++ Source/Game/Entities/Creature/Creature.cs
@@ -2213,7 +2213,7 @@ namespace Game.Entities
                 return;
             }
 
-            uint thisRespawnTime = (uint)(forceDelay != 0 ? Time.UnixTime + forceDelay : m_respawnTime);
+            long thisRespawnTime = forceDelay != 0 ? Time.UnixTime + forceDelay : m_respawnTime;
             GetMap().SaveRespawnTime(SpawnObjectType.Creature, m_spawnId, GetEntry(), thisRespawnTime, GetMap().GetZoneId(GetPhaseShift(), GetHomePosition()), GridDefines.ComputeGridCoord(GetHomePosition().GetPositionX(), GetHomePosition().GetPositionY()).GetId(), saveToDb && m_creatureData != null && m_creatureData.dbData);
         }
